pyproject.toml
[project]
name = "django-api-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "django==5.2",
    "django-crispy-forms==2.4",
    "django-environ>=0.13.0",
    "djangorestframework>=3.17.1",
    "dotenv>=0.9.9",
    "pillow>=12.0.0",
    "sqlparse==0.5.3",
    "tzdata==2025.2",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.0.0",
]
